+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 4.2.12
+
+- update tidy3d client to latest version 1.0.2
+- add `is_3d` to run simulations in 2D
+
 ## 4.2.11
 
 - tidy3d simulation plugin improvements